Start of an iPhone GUI
I built a basic screen with a text box and translate button just to get
us started on building this as an app. I haven't yet figured out how to
connect the visuals to any kind of action, maybe someone else has more
experience with that. Anyways, it works on the iPhone emulator and
we'll have something visual to show to Hamm in addition to the text
based morse code translating we already have.